Fixed so many bugs
you can't create a list from the number the user inputs. You need to create the list of the string
I tested the length and decimal they both work pretty well except that if you have 0s at the right of the decimal as your ending numbers those functions are not going to count them. For ex. the decimal function returned 2 for 0.2100000 and the length function returned 3 for 3.1400000
I added line 56-58  you can go read the comment
The sci_note function has a small visual error that needs to be fixed (I will do it someday)
I don't know why length and decimals are working since I looked through the code and I couldn't figure it out either